+++ apps/frontend/src/app/5_pages/ProtocolDataPage/ProtocolDataPage.tsx
@@ -5,13 +5,9 @@ import { Helmet } from 'react-helmet-async';
 
 import { Heading, Paragraph, ParagraphSize } from '@sovryn/ui';
 
-import { RSK_CHAIN_ID } from '../../../config/chains';
-
-import { NetworkBanner } from '../../2_molecules/NetworkBanner/NetworkBanner';
-import {
-  SECTION_LINKS_LABELS,
-  pageTranslations,
-} from './ProtocolDataPage.constants';
+import { useCurrentChain } from '../../../hooks/useChainStore';
+import { isBobChain } from '../../../utils/chain';
+import { pageTranslations } from './ProtocolDataPage.constants';
 import { EcosystemStatistics } from './components/EcosystemStatistics/EcosystemStatistics';
 import { LendAndBorrow } from './components/LendAndBorrow/LendAndBorrow';
 import { MarketMaking } from './components/MarketMaking/MarketMaking';
@@ -19,37 +15,42 @@ import { PricesBy24hChange } from './components/PricesBy24hChange/PricesBy24hCha
 import { SectionLinks } from './components/SectionLinks/SectionLinks';
 
 const ProtocolDataPage: FC = () => {
-  const refs = useMemo(
-    () => ({
-      ecosystemStatisticsRef: createRef<HTMLDivElement>(),
-      pricesBy24hChangeRef: createRef<HTMLDivElement>(),
-      lendAndBorrowRef: createRef<HTMLDivElement>(),
-      marketMakingRef: createRef<HTMLDivElement>(),
-    }),
-    [],
-  );
+  const chainId = useCurrentChain();
 
-  const sections = useMemo(
-    () => [
+  const sections = useMemo(() => {
+    if (isBobChain(chainId)) {
+      return [
+        {
+          ref: createRef<HTMLDivElement>(),
+          content: <EcosystemStatistics />,
+          title: t(pageTranslations.sections.ecosystemStatistics),
+        },
+      ];
+    }
+
+    return [
       {
-        ref: refs.pricesBy24hChangeRef,
+        ref: createRef<HTMLDivElement>(),
         content: <PricesBy24hChange />,
+        title: t(pageTranslations.pricesBy24hChange.title),
       },
       {
-        ref: refs.lendAndBorrowRef,
+        ref: createRef<HTMLDivElement>(),
         content: <LendAndBorrow />,
+        title: t(pageTranslations.lendAndBorrow.title),
       },
       {
-        ref: refs.marketMakingRef,
+        ref: createRef<HTMLDivElement>(),
         content: <MarketMaking />,
+        title: t(pageTranslations.marketMaking.title),
       },
       {
-        ref: refs.ecosystemStatisticsRef,
+        ref: createRef<HTMLDivElement>(),
         content: <EcosystemStatistics />,
+        title: t(pageTranslations.sections.ecosystemStatistics),
       },
-    ],
-    [refs],
-  );
+    ];
+  }, [chainId]);
 
   return (
     <>
@@ -58,32 +59,27 @@ const ProtocolDataPage: FC = () => {
       </Helmet>
 
       <div className="w-full  text-gray-10 md:mx-9 mx-0 md:mb-2 md:mt-7 mt-4 mb-7 max-w-6xl">
-        <NetworkBanner
-          requiredChainId={RSK_CHAIN_ID}
-          childClassName="grid grid-cols-1 justify-center"
+        <Heading className="text-center mb-3 lg:text-2xl text-base font-medium">
+          {t(pageTranslations.title)}
+        </Heading>
+
+        <Paragraph
+          className="text-center mb-6 lg:mb-9 font-semibold px-4 sm:px-0"
+          size={ParagraphSize.base}
         >
-          <Heading className="text-center mb-3 lg:text-2xl text-base font-medium">
-            {t(pageTranslations.title)}
-          </Heading>
+          {t(pageTranslations.subtitle)}
+        </Paragraph>
 
-          <Paragraph
-            className="text-center mb-6 lg:mb-9 font-semibold px-4 sm:px-0"
-            size={ParagraphSize.base}
-          >
-            {t(pageTranslations.subtitle)}
-          </Paragraph>
+        <SectionLinks
+          refs={sections.map(section => section.ref)}
+          labels={sections.map(section => section.title)}
+        />
 
-          <SectionLinks
-            refs={Object.values(refs)}
-            labels={SECTION_LINKS_LABELS}
-          />
-
-          {sections.map((section, index) => (
-            <div key={index} ref={section.ref}>
-              {section.content}
-            </div>
-          ))}
-        </NetworkBanner>
+        {sections.map((section, index) => (
+          <div key={index} ref={section.ref}>
+            {section.content}
+          </div>
+        ))}
       </div>
     </>
   );