+++ apps/frontend/src/app/5_pages/ProtocolDataPage/components/EcosystemStatistics/EcosystemStatistics.types.ts
@@ -1,11 +1,12 @@
+import { ReactNode } from 'react';
+
 export type ContractData = {
   title: string;
-  value: number;
-  dataAttribute: string;
+  value: ReactNode;
   highlight?: boolean;
 };
 
 export type ContractDataItem = {
-  totalBtc: number;
-  totalUsd: number;
+  totalBtc: string;
+  totalUsd: string;
 };