+++ apps/frontend/src/utils/indexer.ts
@@ -1,24 +1,20 @@
 import { ChainId } from '@sovryn/ethers-provider';
 
+import { SOVRYN_INDEXER } from '../constants/infrastructure';
 import { BOB } from '../constants/infrastructure/bob';
 import { SEPOLIA } from '../constants/infrastructure/sepolia';
 import { Environments } from '../types/global';
-import { isBobChain } from './chain';
+import { isBobChain, isRskChain } from './chain';
 import { isMainnet } from './helpers';
 
-export const getIndexerUri = (chainId: ChainId) => {
-  if (isBobChain(chainId)) {
-    return BOB.indexer[
+export const getIndexerUri = (chainId: ChainId) =>
+  BOB.indexer[isMainnet() ? Environments.Mainnet : Environments.Testnet];
+
+export const getSovrynIndexerUri = (chainId: ChainId) => {
+  if (isBobChain(chainId) || isRskChain(chainId)) {
+    return SOVRYN_INDEXER[
       isMainnet() ? Environments.Mainnet : Environments.Testnet
     ];
-  }
-
-  return SEPOLIA.indexer.testnet;
-};
-
-export const getSdexUri = (chainId: ChainId) => {
-  if (isBobChain(chainId)) {
-    return BOB.sdex[isMainnet() ? Environments.Mainnet : Environments.Testnet];
   }
 
   return SEPOLIA.indexer.testnet;