+++ packages/ui/src/1_atoms/index.tsx
@@ -14,5 +14,6 @@ export * from './DynamicValue';
 export * from './Lottie';
 export * from './ErrorBadge';
 export * from './Toggle';
+export * from './Slider';
 export * from './Bar/Bar';
 export * from './Slider/Slider';