+++ packages/ui/src/2_molecules/Tooltip/Tooltip.tsx
@@ -23,6 +23,7 @@ import {
   TooltipPlacement,
   TooltipTrigger,
   TooltipEvents,
+  TooltipStyle,
 } from './Tooltip.types';
 import { CLOSE_DELAY, getTooltipPosition } from './Tooltip.utils';
 
@@ -41,6 +42,7 @@ type TooltipProps = {
   onHide?: () => void;
   disabled?: boolean;
   trigger?: TooltipTrigger;
+  style?: TooltipStyle;
 };
 
 export const Tooltip: FC<TooltipProps> = ({
@@ -55,6 +57,7 @@ export const Tooltip: FC<TooltipProps> = ({
   onHide,
   disabled = false,
   trigger = TooltipTrigger.hover,
+  style = TooltipStyle.secondary,
 }) => {
   const tooltipRef = useRef<HTMLDivElement>(null);
   const targetRef = useRef<HTMLElement>(null);
@@ -177,6 +180,7 @@ export const Tooltip: FC<TooltipProps> = ({
               styles.tooltip,
               styles[tooltipPosition?.arrowStyles],
               tooltipClassName,
+              styles[style],
             )}
             style={tooltipPosition?.positionStyles}
             ref={tooltipRef}